+++ src/Transaction/Transaction.php
@@ -30,54 +30,9 @@
  * Please do not use the plugin if you do not agree to these terms of use!
  */
 
-namespace Wirecard\PaymentSdk;
+namespace Wirecard\PaymentSdk\Transaction;
 
-/**
- * Class CreditCard3DTransaction
- * @package Wirecard\PaymentSdk
- *
- * An immutable entity representing a 3D payment with a credit card.
- */
-class ThreeDCreditCardTransaction extends CreditCardTransaction
+interface Transaction
 {
-    /**
-     * @var string
-     */
-    private $notificationUrl;
 
-    /**
-     * @var string
-     */
-    private $termUrl;
-
-    /**
-     * ThreeDCreditCardTransaction constructor.
-     * @param Money $money
-     * @param $tokenId
-     * @param $notificationUrl
-     * @param $termUrl
-     */
-    public function __construct($money, $tokenId, $notificationUrl, $termUrl)
-    {
-        parent::__construct($money);
-        $this->setTokenId($tokenId);
-        $this->notificationUrl = $notificationUrl;
-        $this->termUrl = $termUrl;
-    }
-
-    /**
-     * @return string
-     */
-    public function getNotificationUrl()
-    {
-        return $this->notificationUrl;
-    }
-
-    /**
-     * @return string
-     */
-    public function getTermUrl()
-    {
-        return $this->termUrl;
-    }
 }
